Refactor parallelization infrastructure
