+++ algorithms/src/fft/domain.rs
@@ -26,11 +26,16 @@
 //! This allows us to perform polynomial operations in O(n)
 //! by performing an O(n log n) FFT over such a domain.
 
-use crate::fft::{DomainCoeff, SparsePolynomial};
+use crate::{
+    cfg_chunks_mut,
+    cfg_into_iter,
+    cfg_iter,
+    cfg_iter_mut,
+    fft::{DomainCoeff, SparsePolynomial},
+};
 use snarkvm_fields::{batch_inversion, FftField, FftParameters, Field};
-use snarkvm_utilities::{errors::SerializationError, serialize::*};
+use snarkvm_utilities::{errors::SerializationError, execute_with_max_available_threads, serialize::*};
 
-use aleo_std::prelude::*;
 use rand::Rng;
 use std::fmt;
 
@@ -160,7 +165,7 @@ impl<F: FftField> EvaluationDomain<F> {
 
     /// Compute an FFT, modifying the vector in place.
     pub fn fft_in_place<T: DomainCoeff<F>>(&self, coeffs: &mut Vec<T>) {
-        execute_in_threadpool(|| {
+        execute_with_max_available_threads(|| {
             coeffs.resize(self.size(), T::zero());
             self.in_order_fft_in_place(&mut *coeffs);
         });
@@ -176,7 +181,7 @@ impl<F: FftField> EvaluationDomain<F> {
     /// Compute an IFFT, modifying the vector in place.
     #[inline]
     pub fn ifft_in_place<T: DomainCoeff<F>>(&self, evals: &mut Vec<T>) {
-        execute_in_threadpool(|| {
+        execute_with_max_available_threads(|| {
             evals.resize(self.size(), T::zero());
             self.in_order_ifft_in_place(&mut *evals);
         });
@@ -192,7 +197,7 @@ impl<F: FftField> EvaluationDomain<F> {
     /// Compute an FFT over a coset of the domain, modifying the input vector
     /// in place.
     pub fn coset_fft_in_place<T: DomainCoeff<F>>(&self, coeffs: &mut Vec<T>) {
-        execute_in_threadpool(|| {
+        execute_with_max_available_threads(|| {
             Self::distribute_powers(coeffs, F::multiplicative_generator());
             self.fft_in_place(coeffs);
         });
@@ -207,7 +212,7 @@ impl<F: FftField> EvaluationDomain<F> {
 
     /// Compute an IFFT over a coset of the domain, modifying the input vector in place.
     pub fn coset_ifft_in_place<T: DomainCoeff<F>>(&self, evals: &mut Vec<T>) {
-        execute_in_threadpool(|| {
+        execute_with_max_available_threads(|| {
             evals.resize(self.size(), T::zero());
             self.in_order_coset_ifft_in_place(&mut *evals);
         });
@@ -356,31 +361,6 @@ impl<F: FftField> EvaluationDomain<F> {
         cfg_iter_mut!(result).zip(other_evals).for_each(|(a, b)| *a *= b);
 
         result
-    }
-}
-
-#[inline(always)]
-fn execute_in_threadpool(f: impl FnOnce() + Send) {
-    #[cfg(feature = "parallel")]
-    {
-        let num_threads = match aleo_std::get_cpu() {
-            Cpu::Intel | Cpu::Unknown => {
-                let rayon_threads = rayon::current_num_threads();
-                num_cpus::get_physical().min(rayon_threads)
-            }
-            Cpu::AMD => rayon::current_num_threads(),
-        };
-
-        let pool = rayon::ThreadPoolBuilder::new()
-            .num_threads(num_threads)
-            .build()
-            .unwrap();
-
-        pool.install(f);
-    }
-    #[cfg(not(feature = "parallel"))]
-    {
-        f();
     }
 }
