+++ utilities/Cargo.toml
@@ -22,11 +22,22 @@ path = "../derives"
 version = "0.7.5"
 optional = true
 
+[dependencies.aleo-std]
+version = "0.1.8"
+optional = true
+
 [dependencies.anyhow]
 version = "1.0"
 
 [dependencies.bincode]
 version = "1.3.3"
+
+[dependencies.itertools]
+version = "0.10.3"
+
+[dependencies.num_cpus]
+version = "1"
+optional = true
 
 [dependencies.num-bigint]
 version = "0.4"
@@ -35,6 +46,10 @@ version = "0.4"
 version = "0.8"
 default-features = false
 features = [ "std_rng" ]
+
+[dependencies.rayon]
+version = "1"
+optional = true
 
 [dependencies.serde]
 version = "1.0"
@@ -51,3 +66,4 @@ default-features = false
 default = [ "std", "derive" ]
 std = [ ]
 derive = [ "snarkvm-derives" ]
+parallel = [ "rayon", "num_cpus", "aleo-std" ]
